backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.4.0
PyJWT==2.8.0
PyMySQL==1.1.1
psycopg2-binary==2.9.9
numpy==2.4.3
gunicorn==21.2.0
python-dotenv==1.0.0
pypdf==4.3.1
hypothesis==6.151.12
pytest==9.0.1
pytest-django==4.12.0
